requirements.txt
Django~=4.0
pytest~=6.2.5
djangorestframework~=3.13.1
boto3~=1.24.46
celery~=5.2.3
botocore~=1.27.46
django-filter~=21.1
pyparsing~=3.0.6
cryptography~=37.0.2
python-dotenv~=0.19.2
Jinja2~=3.1.2